resource-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pyjwt==2.8.0
pydantic==2.7.1
